+++ src/Loaders/ProductValueLoader.php
@@ -52,6 +52,11 @@ class ProductValueLoader implements LoaderInterface
     protected $registryLoader;
 
     /**
+     * @var bool
+     */
+    private $loaded = false;
+
+    /**
      * Construct that initializes the iterator with the product processor instance.
      *
      * @param \TechDivision\Import\Loaders\LoaderInterface                         $registryLoader   The registry loader instance
@@ -78,12 +83,17 @@ class ProductValueLoader implements LoaderInterface
      */
     public function load()
     {
+        if ($this->loaded) {
+            // return the values
+            return $this->values;
+        }
+
         // load the already processed SKUs from the registry, merge
         // them with the ones from the DB and return them
         $collectedColumns = $this->getRegistryLoader()->load();
 
         // query whether or not values for the configured column name are available
-        if (is_array($collectedColumns[$this->columnName])) {
+        if (isset($collectedColumns[$this->columnName]) && is_array($collectedColumns[$this->columnName])) {
             // if yes merge the values into the array with the values from the DB
             foreach ($collectedColumns[$this->columnName] as $value) {
                 // query whether or not the value already exits
@@ -96,6 +106,8 @@ class ProductValueLoader implements LoaderInterface
             }
         }
 
+        // all SKUs loaded from DB and import files
+        $this->loaded = true;
         // return the values
         return $this->values;
     }